+++ src/om_bootstrap/mixins.cljs
@@ -44,3 +44,53 @@
   (set-timeout [owner f timeout]
                (let [timeout (js/setTimeout f timeout)]
                  (.push (.-timeouts owner) timeout))))
+
+;; ## Dropdown Mixin
+
+(defn in-root?
+  "Accepts two DOM elements; returns true if the supplied node is
+  nested inside the supplied root, false otherwise."
+  [node root]
+  (loop [node node]
+    (cond (nil? node) false
+          (= root node) true
+          :else (recur (.-parentNode node)))))
+
+(defn bind-root-close-handlers!
+  "For dropdowns, binds a handler for the click and "
+  [owner]
+  (let [set-state (aget owner "setDropdownState")]
+    (set! (.-onDocumentClickListener owner)
+                (event-listener
+                 js/document "click"
+                 (fn [e]
+                   (when-not (in-root? (.-target e) (om/get-node owner))
+                     (set-state false)))))
+    (set! (.-onDocumentKeyupListener owner)
+          (event-listener
+           js/document "keyup"
+           (fn [e]
+             (when (= 27 (.-keyCode e))
+               (set-state false)))))))
+
+(defn unbind-root-close-handlers! [owner]
+  (when-let [l (.-onDocumentClickListener owner)]
+    (l)
+    (set! (.-onDocumentKeyupListener owner) nil))
+  (when-let [l (.-onDocumentKeyupListener owner)]
+    (l)
+    (set! (.-onDocumentKeyupListener owner) nil)))
+
+(defmixin dropdown-mixin
+  "Note that we're going to have to do some magic to get the state
+  setting to work
+  here. https://github.com/Prismatic/om-tools/issues/28"
+  (init-state [_] {:open? false})
+  (will-unmount [owner] (unbind-root-close-handlers! owner))
+  (isDropdownOpen [owner] (om/get-state owner :open?))
+  (setDropdownState
+   [owner open?]
+   (if open?
+     (bind-root-close-handlers! owner)
+     (unbind-root-close-handlers! owner))
+   (om/set-state! owner [:open?] open?)))